Fixed stuff that needed to be fixed
