+++ requirements.txt
@@ -2,4 +2,5 @@ termcolor==1.1.0
 Keras==2.2.4
 numpy==1.16.3
 Flask==1.0.2
-scikit_learn==0.21.3
+scikit-learn==0.21.3
+tensorflow
